+++ src/utils.ts
@@ -1,5 +1,9 @@
 import { DataFrame, FieldType, LinkModel } from '@grafana/data';
 import { findField } from '@volkovlabs/grafana-utils';
+import { Base64 } from 'js-base64';
+
+import { BASE64_MEDIA_HEADER_REGEX, IMAGE_TYPES_SYMBOLS } from './constants';
+import { SupportedFileType } from './types';
 
 /**
  * Convert Base64 to Blob
@@ -44,4 +48,41 @@ export const getDataLink = (frames: DataFrame[], optionName: string, currentInde
   }
 
   return null;
+};
+
+/**
+ * Handle media data
+ * @param mediaField
+ */
+export const handleMediaData = (mediaField: string | undefined) => {
+  let currentMedia = mediaField;
+  let type;
+
+  if (mediaField) {
+    const mediaMatch = mediaField.match(BASE64_MEDIA_HEADER_REGEX);
+
+    if (!mediaMatch?.length) {
+      /**
+       * Encode to base64 if not
+       */
+
+      if (!Base64.isValid(mediaField)) {
+        currentMedia = Base64.encode(mediaField);
+      }
+
+      /**
+       * Set header
+       */
+      type = IMAGE_TYPES_SYMBOLS[mediaField.charAt(0)];
+
+      currentMedia = type ? `data:${type};base64,${currentMedia}` : `data:;base64,${currentMedia}`;
+    } else if (Object.values(SupportedFileType).includes(mediaMatch[1] as SupportedFileType)) {
+      type = mediaMatch[1];
+    }
+  }
+
+  return {
+    currentMedia,
+    type,
+  };
 };